+++ package.json
@@ -16,8 +16,8 @@
     "prebuild": "yarn run clean",
     "clean": "del dist",
     "prepublishOnly": "yarn run build",
-    "pretest": "yarn run build",
-    "test": "ava -v && standard"
+    "test": "ava -v && standard",
+    "seed-test": "ava --match='*seed*'"
   },
   "engines": {
     "node": ">=10"
@@ -43,6 +43,7 @@
     "bates"
   ],
   "devDependencies": {
+    "@ava/typescript": "^1.1.1",
     "@types/node": "^14.14.31",
     "@types/seedrandom": "^2.4.28",
     "ava": "^3.5.0",
@@ -52,6 +53,7 @@
     "dev-cli": "^0.0.0",
     "esm": "^3.2.25",
     "standard": "^14.3.3",
+    "ts-node": "^9.1.1",
     "tsc": "^1.20150623.0",
     "typescript": "^4.1.5",
     "update-markdown-jsdoc": "^1.0.11"
@@ -67,8 +69,11 @@
     "dist"
   ],
   "ava": {
+    "extensions": [
+			"ts"
+		],
     "require": [
-      "esm"
+      "ts-node/register"
     ]
   },
   "babel": {
