--- test/_in-delta.js
@@ -1,3 +1,0 @@
-export default (actual, expected, delta) => (
-  expected - delta <= actual && actual <= expected + delta
-)
