+++ test/_in-delta.ts
@@ -1,0 +1,3 @@
+export default (actual: number, expected: number, delta: number) => (
+  expected - delta <= actual && actual <= expected + delta
+)
