+++ test/distributions/irwinHall.test.ts
@@ -1,11 +1,11 @@
 import test from 'ava'
 import seedrandom from 'seedrandom'
 
-import random from '../..'
+import random from "../../src/random"
 
-test('random.logNormal() produces numbers', (t) => {
+test('random.irwinHall() produces numbers', (t) => {
   const r = random.clone(seedrandom('ZDJjM2IyNmFlNmVjNWQwMGZkMmY1Y2Nk'))
-  const d = r.logNormal()
+  const d = r.irwinHall()
   for (let i = 0; i < 10000; ++i) {
     const v = d()
     t.is(typeof v, 'number')
