+++ test/distributions/uniform.test.ts
@@ -2,7 +2,7 @@ import test from 'ava'
 import seedrandom from 'seedrandom'
 
 import inDelta from '../_in-delta'
-import random from '../..'
+import random from "../../src/random"
 
 test('random.uniform() is in [0, 1)', (t) => {
   const r = random.clone(seedrandom('ZDJjM2IyNmFlNmVjNWQwMGZkMmY1Y2Nk'))
