+++ test/seed.test.ts
@@ -1,9 +1,9 @@
 import test from 'ava'
 import seedrandom from 'seedrandom'
 
-import random from '..'
+import random from "../src/random"
 
-test('random.clone witth a seed is consistent', (t) => {
+test('random.clone with a seed is consistent', (t) => {
   const r = random.clone(seedrandom('ZjExZDczNWQxY2NlZjUzYmRiZWU0ZGIz'))
   const d = r.uniform()
   const o = []
