+++ test/snapshots/seed.test.ts.md
@@ -1,0 +1,112 @@
+# Snapshot report for `test/seed.test.ts`
+
+The actual snapshot is saved in `seed.test.ts.snap`.
+
+Generated by [AVA](https://avajs.dev).
+
+## random.clone with a seed is consistent
+
+> Snapshot 1
+
+    [
+      0.9362808130793461,
+      0.4864517867266793,
+      0.5065797249862649,
+      0.3300815411147545,
+      0.022876541849241913,
+      0.0018718398576695902,
+      0.8458705577357042,
+      0.9501863429937075,
+      0.5562872420975038,
+      0.8032030618791965,
+      0.35937030051611374,
+      0.5917219992006454,
+      0.4325656059975954,
+      0.9379514250312311,
+      0.46913700253375373,
+      0.04152165460184985,
+      0.10489461372302518,
+      0.39999289745341915,
+      0.8801355439468278,
+      0.09222709887818471,
+      0.06597322766572979,
+      0.1015894588008499,
+      0.3657280331015485,
+      0.6011111104408471,
+      0.78326449381139,
+      0.59420855175205,
+      0.053691966139575976,
+      0.4158817510889663,
+      0.00931570708166847,
+      0.9701229323500713,
+      0.004717584311972597,
+      0.19287729568586956,
+      0.6181548641692531,
+      0.37000120506984135,
+      0.6005060662035266,
+      0.8122881262971489,
+      0.35463676725026466,
+      0.7967726657451469,
+      0.46183388516953455,
+      0.11933085399009244,
+      0.3545328926842683,
+      0.6508090363525306,
+      0.10107431393274013,
+      0.9714515644272922,
+      0.11151788408134233,
+      0.464713458854543,
+      0.4553200439243976,
+      0.48422476911952295,
+      0.39956302553141226,
+      0.37041811110053097,
+      0.7812664270723839,
+      0.8202762819862192,
+      0.877299251777231,
+      0.029885604135202436,
+      0.408590659878165,
+      0.8702831027886333,
+      0.9964762844416561,
+      0.06278663019936892,
+      0.004478785353696548,
+      0.1616154066206868,
+      0.7699198474818526,
+      0.19881279983122438,
+      0.05653706710280909,
+      0.43287061849854586,
+      0.961911350663349,
+      0.22504486706570723,
+      0.6182062667952513,
+      0.4021421107923509,
+      0.8257999453586723,
+      0.45237237606685804,
+      0.9981277617262307,
+      0.7470690286400854,
+      0.43235970744895824,
+      0.9040477390410735,
+      0.7239179552060097,
+      0.7380685372803176,
+      0.5128163041953872,
+      0.3005013950261629,
+      0.0031608054443676378,
+      0.46498685343066454,
+      0.999019326323084,
+      0.890497641236873,
+      0.8106073356352089,
+      0.737468157432021,
+      0.7148898472029561,
+      0.7674900213546753,
+      0.30781674686227095,
+      0.8198454353724389,
+      0.9256466066420532,
+      0.3794669665753787,
+      0.5431519619824302,
+      0.26530751329149066,
+      0.9761499504404662,
+      0.5066022972533559,
+      0.13506428822435051,
+      0.35358863466610246,
+      0.38264476224040667,
+      0.4574017110975189,
+      0.2782995604601085,
+      0.7622410349799561,
+    ]
